requirements.txt
# ==============================================================================
# PYTHON DEPENDENCIES FOR GCP CLOUD RUN + GCS + BIGQUERY
# ==============================================================================
#
# flask: Web framework for webhook HTTP requests
# gunicorn: Production web server
# google-cloud-bigquery: Write/query data in BigQuery
# google-cloud-storage: Write raw JSON to GCS bucket
# requests: Call Zoom API (for camera data)

flask==3.0.0
gunicorn==21.2.0
google-cloud-bigquery==3.14.1
google-cloud-storage==2.14.0
requests==2.31.0
